[Feature][Jeff] Staff view and dashboard. Staff account added
